Split whole old TTS implementations to a separate modules
